Model_Serving/requirements.txt
bert-for-tf2==0.14.7
Flask==1.1.2
h5py==3.1.0
importlib-metadata==3.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.19.5
params-flow==0.8.2
protobuf==3.14.0
py-params==0.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.11.0
sentencepiece==0.1.95
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
